modules/lambda/src_pptx_modifier/requirements.txt
python-pptx
lxml
Pillow
XlsxWriter
fastapi==0.115.8
mangum==0.19.0
pydantic==2.10.6
PyJWT